update(structure):add gemini service and remove business logic from controller
